docs: 调整了scrollbar\slider\axis的example组织 (#85)
