+++ examples/chart-ui/axis/demo/label-formatter.ts
